Replace emoji-toolkit with combination of skin-tone & emoji-regex
Seems that the toolkit has problems with conversion of some of the
emojis to shortcodes. Plus, I learned that those shortcodes
are not unified so it doesn't make sense to use them.

I decided to go back to skin-tone to generate emojis w/ skin tone
and emoji-regex for testing purpose.

With this commit the shortcode() method is removed.